bpmn-visualization-hackathon/stt/requirements.txt
Flask==2.3.2
openai-whisper==20240930
torch==2.6.0 --index-url https://download.pytorch.org/whl/cpu
ffmpeg-python==0.2.0
numpy==1.26.2
gunicorn==23.0.0